requirements.txt
fastapi
uvicorn
pyodbc
pydantic
python-dotenv
python-socketio
python-socketio[asyncio_client]
python-socketio[asgi]
azure-cognitiveservices-speech
azure-storage-blob
pandas
openpyxl
httpx
librosa
numpy
opencv-contrib-python
opencv-python
torch
torchvision
tqdm
numba